+++ 리아-이아현/2주차/mission01/src/App.jsx
@@ -1,43 +1,23 @@
-import { useState } from "react";
+import { useContext } from "react";
 import "./App.css";
 import Button from "./components/Button";
 import Input from "./components/Input";
+import { TodoContext } from "./context/TodoContext";
 
 function App() {
-  const [todos, setTodos] = useState([{ id: 1, task: "투두 만들어보기" }]);
-  const [text, setText] = useState("");
-  const [editingId, setEditingId] = useState("");
-  const [editText, setEditText] = useState("");
-
-  // 렌더링 방지
-  const handleSubmit = (e) => {
-    e.preventDefault();
-  };
-
-  // 1. 추가하기
-  const addTodo = () => {
-    if (text.trim().length === 0) {
-      alert("할 일을 입력하세요");
-    }
-    setTodos((prev) => [
-      ...prev,
-      { id: Math.floor(Math.random() * 100) + 2, task: text },
-    ]);
-    setText("");
-  };
-
-  // 2. 삭제하기
-  const deleteTodo = (id) => {
-    setTodos((prev) => prev.filter((item) => item.id !== id));
-  };
-
-  // 3. 수정하기
-  const updateTodo = (id, text) => {
-    setTodos((prev) =>
-      prev.map((item) => (item.id === id ? { ...item, task: text } : item))
-    );
-    setEditingId("");
-  };
+  const {
+    todos,
+    text,
+    setText,
+    editingId,
+    setEditingId,
+    editText,
+    setEditText,
+    handleSubmit,
+    addTodo,
+    deleteTodo,
+    updateTodo,
+  } = useContext(TodoContext);
 
   return (
     <>
