+++ 리아-이아현/2주차/mission01/src/context/TodoContext.jsx
@@ -1,0 +1,57 @@
+import { createContext, useState } from "react";
+
+export const TodoContext = createContext();
+
+export function TodoContextProvider({ children }) {
+  const [todos, setTodos] = useState([{ id: 1, task: "투두 만들어보기" }]);
+  const [text, setText] = useState("");
+  const [editingId, setEditingId] = useState("");
+  const [editText, setEditText] = useState("");
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+  };
+
+  const addTodo = () => {
+    if (text.trim().length === 0) {
+      alert("할 일을 입력하세요");
+    }
+    setTodos((prev) => [
+      ...prev,
+      { id: Math.floor(Math.random() * 100) + 2, task: text },
+    ]);
+    setText("");
+  };
+
+  const deleteTodo = (id) => {
+    setTodos((prev) => prev.filter((item) => item.id !== id));
+  };
+
+  const updateTodo = (id, text) => {
+    setTodos((prev) =>
+      prev.map((item) => (item.id === id ? { ...item, task: text } : item))
+    );
+    setEditingId("");
+  };
+
+  return (
+    <TodoContext.Provider
+      value={{
+        todos,
+        setTodos,
+        text,
+        setText,
+        editingId,
+        setEditingId,
+        editText,
+        setEditText,
+        handleSubmit,
+        addTodo,
+        deleteTodo,
+        updateTodo,
+      }}
+    >
+      {children}
+    </TodoContext.Provider>
+  );
+}
