+++ 리아-이아현/2주차/mission01/src/main.jsx
@@ -1,7 +1,10 @@
-import { createRoot } from 'react-dom/client'
-import App from './App.jsx'
-import './index.css'
+import { createRoot } from "react-dom/client";
+import App from "./App.jsx";
+import "./index.css";
+import { TodoContextProvider } from "./context/TodoContext.jsx";
 
-createRoot(document.getElementById('root')).render(
+createRoot(document.getElementById("root")).render(
+  <TodoContextProvider>
     <App />
-)
+  </TodoContextProvider>
+);
